Implemented SchemaIOProvider for DataStoreV1, refactored tests
